closes #17, uses pundit in leagues_controller

+++ app/controllers/leagues_controller.rb
@@ -3,13 +3,16 @@ class LeaguesController < ApplicationController
 
   def show
     @league = League.find(params[:id])
+    authorize @league
   end
 
   def new
+    authorize League
     @league = current_user.leagues.new
   end
 
   def create
+    authorize League
     @league = current_user.leagues.new(_league_params)
     if @league.save
       redirect_to @league
@@ -20,6 +23,7 @@ class LeaguesController < ApplicationController
 
   def destroy
     league = League.find(params[:id])
+    authorize league
     league.destroy
     redirect_to dashboard_path
   end
